detailsScreen

+++ TestOnlineShop/app/src/main/java/com/skillbox/skillbox/testonlineshop/classes/Product.kt
@@ -15,7 +15,7 @@ data class Product(
     @SerializedName("title")
     val title: String,
     @SerializedName("rating")
-    val rating: Double?,
+    val rating: Float?,
     @SerializedName("CPU")
     val cpu: String?,
     @SerializedName("camera")
